(PC-35630)[PRO] feat: move all preview in the same folder
