+++ pro/src/components/ImageUploader/components/ButtonImageEdit/ModalImageEdit/components/ModalImageUploadConfirm/ModalImageUploadConfirm.tsx
@@ -2,8 +2,8 @@ import * as Dialog from '@radix-ui/react-dialog'
 
 import { useAnalytics } from 'app/App/analytics/firebase'
 import { Events } from 'commons/core/FirebaseEvents/constants'
-import { AppPreviewOffer } from 'components/ImageUploader/components/ButtonAppPreview/components/AppPreviewOffer/AppPreviewOffer'
-import { AppPreviewVenue } from 'components/ImageUploader/components/ButtonAppPreview/components/AppPreviewVenue/AppPreviewVenue'
+import { AppPreviewOffer } from 'components/ImageUploader/components/ImagePreview/components/AppPreviewOffer/AppPreviewOffer'
+import { AppPreviewVenue } from 'components/ImageUploader/components/ImagePreview/components/AppPreviewVenue/AppPreviewVenue'
 import { UploaderModeEnum } from 'components/ImageUploader/types'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
