+++ pro/src/components/ImageUploader/components/ImagePreview/ImagePreview.tsx
