+++ pro/src/components/ImageUploader/components/ImagePreview/components/AppPreviewVenue/__specs__/AppPreviewVenue.spec.tsx
