+++ pro/src/components/ImageUploader/components/ImagePreview/stories/HomeScreenShell.module.scss
