+++ pro/src/components/ImageUploader/components/ImagePreview/stories/offer-home-shell.png
