+++ pro/src/components/ImageUploader/components/ImagePreview/stories/offer-shell.png
